ENH: Add rechunker in context manager
Using `create_optimized_zarr_store` as a context
manager allow the usual icclim.index output to
remain the same.
Plus, it takes care automatically of dask spilling
custom configuration for rechunker and cleanup of
intermediary zarr stores.

Having this feature aside also make it easy to
mark it as experimental.