MAINT: Rework rechunking
Following the discussion on S.O[0] it seems
not so optimal to always use a chunking schema
where a dimension is unchunked.

Now instead of passing a `dim` which will be
unchunked, `create_optimized_zarr_store` accepts
a `chunking` dictionary.
The default behavior is unchanged though, it will
still create a zarr store unchunked on time.
This seems handy because the climate indices
relying on daily percentiles will still be much
faster if there is no chunk on "time" because
of the `map_block` used in xclim bootstrapping.py.

[0] https://stackoverflow.com/questions/71570003/why-does-the-amount-of-dask-task-increase-with-an-optimized-chunking-compared