MAINT: Use Registry everywhere
We choose to replace the standard Enum by Registry because Enum has an
API that is confusing.
Indeed, an Enum class describe BOTH each of its values and the catalog
itself.  This is particularly disturbing when enforcing typing.

Using our Registry class we still have a namespace for the container (e.g. FrequencyRegistry)
and we have proper instances of the content (e.g. Frequency).
Note that Registry should not be instantiated, the class is meant to act as a namespace and
an instance of it would not be useful.