FIX[#199]: Make the chunking on DataArrays
When chunking on the Dataset we would try to chunk variables that
have nothing to do with the computation such as the `time_bds`.
This is unnecessary plus if the variable is not properly decoded
and keep a dtype "object" dask would fail to chunk it.
Instead, we chunk the DataArray(s) that we are sure to read for the
computation.