ENH: Add difference_of_means generic indicator
`difference_of_means` can be used to compute the anomaly.
Contrary to the user_index anomaly, difference_of_means
can be used with a single variable and slice_mode is
then used to `groupby` values (for exemple by month).
Note that some slice_mode may not work properly for now.