FIX: consecutive events count with indexer
We were trying to compute the indexer (e.g. date_bound=("01-01", "15-02")
for indices relying on consecutive days (CDD, CSDI...) but it doesn't make much sense.
Indeed, we would try to compute the consecutive days before clipping time, so that it
would take into account spells that would start before or end after the bounds,
but then the bounds are useless! (or at least they are not properly respected)
So now, even on consecutive days indices, we clip time in the pre-processing step.
This is a similar behavior of the `clipped_season` that was introduces in 5.3 but was then removed to clarify the API.