ENH: Auto-convert input precipitations amount to a rate
This is meant to fix the issue we have with e-obs dataset where
the precipitation is not a flux but rather a "thickness_of_rainfall_amount".
Now, it will be automatically converted to a rate using xclim amount2rate function.
Note that if the dataset is discontinuous, amount2rate might convert,
amounts to a wrong rate[0].

[0] https://github.com/Ouranosinc/xclim/pull/1206#discussion_r1000708058