Fix: Re-add is_precipitation_amount check
This is still necessary as xclim hydro context only convert
`mm` to `mm/day` if and only if the input data is a DataArray
and has the proper standard_name.
In icclim, for generic indices, we try to convert
a Threshold unit (bare Pint Unit),
which cannot carry the required standard_name,
so we need to have an automated conversion before.

This fix *avoids* the issue described above by converting
the input DataArray from amount to flux.